+++ NEWS.md
@@ -9,12 +9,19 @@ Towards version 3.1.0: 2023-09
 - added CmdLine::optional_value<T>("-opt"), for options that take a value
   but are optional and do not have a default
 
-- new CmdLine::Result<T>::present() function to tell if an option was
+- added CmdLine::Result<T>::present() function to tell if an option was
   actually present on the command line (e.g. notably for use with
   optional_value results) 
 
 - added CmdLine::start_section(..) CmdLine::end_section() to allow for
   basic sectioning of help text 
+
+- added CmdLine::dump() to return a string with all options and their values,
+  suitable for reading in as an argfile.
+
+- help description gets wrapped to fit in 80 columns, via the new
+  CmdLine::wrap() static function
+
 
 ### other changes
 - -argfile files now allow # as a comment character (in addition to //)
@@ -23,6 +30,9 @@ Towards version 3.1.0: 2023-09
 
 - all string_val, double_val, int_val functions are deprecated. 
   Use value<string>, value<double>, value<int> instead.
+
+- present_or_set(...) is also deprecated. Use optional_value(...) instead
+  and then check if the result is present().
 
 
 version 3.0.0 [undated]